feat: Add Loki and Promtail monitoring applications
This commit adds the necessary YAML files to deploy Loki and Promtail monitoring applications using ArgoCD. The Loki application is sourced from the Grafana Helm charts repository and is set to version 6.9.0. The Promtail application is also sourced from the same repository and is set to version 6.16.4. Both applications are configured to sync automatically, with pruning and self-healing enabled. The GrafanaDatasource resource for Loki is also included in this commit.

Signed-off-by: Aleksey Sviridkin <f@lex.la>